Fix not responding for every fifth-time touch.
Issue: https://github.com/alexandred/VoodooI2C/issues/185

The correcting of the wrapper index in ``handleDigitizerReport()''
is not a good fit for some specific touchpad model, in which
the first identifier is changed for every five times of touch,
resulting in the wrong actual index for the wrapper.

This fix is simple and may not perfect for every model. We still need to dig into it.